Update expressions.md
I think we should enforce the '=>' in match expressions be in a straight column.  This makes code much easier to read, and since we will be using an automatic formatting tool, it's not hard to keep up-to-date (the usual argument against such a style).  I also think the types in a struct definition should be in a straight column.