Update broken link
Toml now uses toml.io for released specifications and the github repo
for development.  Also the old link was for the 0.4 specification, while
cargo uses toml_edit, which uses toml 1.0 (reference:
https://github.com/toml-rs/toml/blob/main/crates/toml_edit/CHANGELOG.md#030---2021-09-13).
Finally, the link was referring to discussion of "Bare keys" vs "Quoted
keys".  This discussion was in the `#table` section in the 0.4
specification, but is under `#keys` in the 1.0 specification.