Fix links
The style guide no longer lives in this repo; change references to the
style guide to point to it in rust-lang/rust.

Change fmt-rfcs to style-team everywhere it appears (except README,
which needs an overhaul).

Change references to the HEAD branch of repositories to use `HEAD`
rather than the name of the branch, to make it easy to rename.